+++ NEWS.md
@@ -3,6 +3,7 @@
 ## Minor update
 * `age_generator()`: add `categorise` argument to generate additional factor variable for age. 
 * `age_generator()`: add `paediatric` argument to generate additional paediatric age variable.
+* args fixed in `age_generator.default()` 
 
 ---